+++ src/ArduinoJson/Deserialization/deserialize.hpp
@@ -24,14 +24,10 @@ struct first_or_void<T, Rest...> {
 
 // A meta-function that returns true if T is a valid destination type for
 // deserialize()
-template <class T, class = void>
-struct is_deserialize_destination : false_type {};
-
 template <class T>
-struct is_deserialize_destination<
-    T, enable_if_t<is_same<decltype(VariantAttorney::getResourceManager(
-                               detail::declval<T&>())),
-                           ResourceManager*>::value>> : true_type {};
+using is_deserialize_destination =
+    bool_constant<is_base_of<JsonDocument, remove_cv_t<T>>::value ||
+                  IsVariant<T>::value>;
 
 template <typename TDestination>
 inline void shrinkJsonDocument(TDestination&) {